laba5/requirements.txt
Django==5.0.1
# Pillow==10.2.0 # Установим отдельно
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
django-filter==23.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3 